Bugfix + Update MinMaxAmplitudeEstimations, ErrorNotifications etc.
New developments
1) calculate and output data for flagellar parameters at each arc length
position only if at this position the parameter could be quantified for
at least 70% of the frames
2) new method to determine minima and maxima in flagellar parameters
(e.g. curvature angle)
previous method:
- minimum = median of 5 lowest peak values across the entire time series
(in this case, peak values were defined as values whose neighbors were
higher than the value)
- maximum = median of 5 highest peak values across the entire time
series (in this case, peak values were defined as values whose neighbors
were lower than the value)
- amplitude = maximum - minimum
new method:
- split the time series into chunks of 3 beat cycles each (the number of
frames corresponding to one beat cycle was calculated as image
acquisiton frequency (= fps) / primary frequency peak for angle theta)
- consider each chunk only if for at least 70% of the the frames in the
chunk the investigated parameter could be determined
- determine the maximum and minimum in each chunk
- maximum = median of all maxima of all considered chunks
- minimum = median of all minima of all considered chunks
- amplitude = maximum - minimum
3) notify exceptions occuring during processing in the log window with
detailed exception information
4) skip files for which errors occur, notify that, and finish analysis
of the error-free files

Fixed bugs existing in previous versions:
A) the average ("_avg") txt.files contained the median but not the
average and the median ("_medi") files contained the average but not the
median.
B) the x axis in frequency plots of z and curvature angle in the PDF was
incorrect
C) the y axis of the y average, curvature angle, and power of z (FFT
analysis) was incorrect in the output PDF
D) software crashed when an analysis result was loaded were the
intensity in the head was not defined for the same amount of time frames
as the angle theta or the velocity
E) power of z: y axis was wrongly labeled (is not in micron but
normalized to minimum and maximum of determined values)
F) if for a loaded file, an error occured during processing, the
software crashes